New Qwt version 6. Remove of zlib headers. Integration of Qt 5 beta 2. Some updates to osdab.

+++ code/external/qwt/headers/qwt_round_scale_draw.h
@@ -35,8 +35,8 @@ public:
     QwtRoundScaleDraw();
     virtual ~QwtRoundScaleDraw();
 
-    void setRadius( double radius );
-    double radius() const;
+    void setRadius( int radius );
+    int radius() const;
 
     void moveCenter( double x, double y );
     void moveCenter( const QPointF & );
